--- a/crates/circuit/src/circuit_data.rs
+++ b/crates/circuit/src/circuit_data.rs
@@ -13,6 +13,7 @@
 use crate::circuit_instruction::{
     convert_py_to_operation_type, operation_type_and_data_to_py, CircuitInstruction,
 };
+use crate::imports::{BUILTIN_LIST, CLBIT, QUBIT};
 use crate::intern_context::{BitType, IndexType, InternContext};
 use crate::operations::{OperationType, Param};
 use crate::parameter_table::{ParamEntry, ParamTable};
@@ -193,16 +194,30 @@ impl CircuitData {
             global_phase,
         };
         if num_qubits > 0 {
-            let qubit_mod = py.import_bound("qiskit.circuit.quantumregister")?;
-            let qubit_cls = qubit_mod.getattr("Qubit")?;
+            let qubit_cls = QUBIT
+                .get_or_init(py, || {
+                    py.import_bound("qiskit.circuit.quantumregister")
+                        .unwrap()
+                        .getattr("Qubit")
+                        .unwrap()
+                        .unbind()
+                })
+                .bind(py);
             for _i in 0..num_qubits {
                 let bit = qubit_cls.call0()?;
                 res.add_qubit(py, &bit, true)?;
             }
         }
         if num_clbits > 0 {
-            let clbit_mod = py.import_bound(intern!(py, "qiskit.circuit.classicalregister"))?;
-            let clbit_cls = clbit_mod.getattr(intern!(py, "Clbit"))?;
+            let clbit_cls = CLBIT
+                .get_or_init(py, || {
+                    py.import_bound("qiskit.circuit.classicalregister")
+                        .unwrap()
+                        .getattr("Clbit")
+                        .unwrap()
+                        .unbind()
+                })
+                .bind(py);
             for _i in 0..num_clbits {
                 let bit = clbit_cls.call0()?;
                 res.add_clbit(py, &bit, true)?;
@@ -289,8 +304,15 @@ impl CircuitData {
                 })
                 .collect();
             if !params.is_empty() {
-                let builtins = PyModule::import_bound(py, "builtins")?;
-                let list_builtin = builtins.getattr("list")?;
+                let list_builtin = BUILTIN_LIST
+                    .get_or_init(py, || {
+                        PyModule::import_bound(py, "builtins")
+                            .unwrap()
+                            .getattr("list")
+                            .unwrap()
+                            .unbind()
+                    })
+                    .bind(py);
 
                 for (param_index, param) in &params {
                     let temp: PyObject = param.getattr(py, intern!(py, "parameters"))?;
@@ -327,8 +349,16 @@ impl CircuitData {
 
     /// Remove an index's entries from the parameter table.
     fn remove_from_parameter_table(&mut self, py: Python, inst_index: usize) -> PyResult<()> {
-        let builtins = PyModule::import_bound(py, "builtins")?;
-        let list_builtin = builtins.getattr(intern!(py, "list"))?;
+        let list_builtin = BUILTIN_LIST
+            .get_or_init(py, || {
+                PyModule::import_bound(py, "builtins")
+                    .unwrap()
+                    .getattr("list")
+                    .unwrap()
+                    .unbind()
+            })
+            .bind(py);
+
         if inst_index == usize::MAX {
             if let Param::ParameterExpression(global_phase) = &self.global_phase {
                 let temp: PyObject = global_phase.getattr(py, intern!(py, "parameters"))?;
@@ -1065,8 +1095,16 @@ impl CircuitData {
 
     #[setter]
     pub fn global_phase(&mut self, py: Python, angle: Param) -> PyResult<()> {
-        let builtins = PyModule::import_bound(py, "builtins")?;
-        let list_builtin = builtins.getattr(intern!(py, "list"))?;
+        let list_builtin = BUILTIN_LIST
+            .get_or_init(py, || {
+                PyModule::import_bound(py, "builtins")
+                    .unwrap()
+                    .getattr("list")
+                    .unwrap()
+                    .unbind()
+            })
+            .bind(py);
+
         self.remove_from_parameter_table(py, usize::MAX)?;
         match angle {
             Param::Float(angle) => {