--- a/crates/circuit/src/circuit_data.rs
+++ b/crates/circuit/src/circuit_data.rs
@@ -589,7 +589,8 @@ impl CircuitData {
     ///
     /// Returns:
     ///     CircuitData: The shallow copy.
-    pub fn copy(&self, py: Python<'_>) -> PyResult<Self> {
+    #[pyo3(signature = (copy_instructions=true))]
+    pub fn copy(&self, py: Python<'_>, copy_instructions: bool) -> PyResult<Self> {
         let mut res = CircuitData::new(
             py,
             Some(self.qubits.bind(py)),
@@ -602,36 +603,38 @@ impl CircuitData {
         res.data.clone_from(&self.data);
         res.param_table.clone_from(&self.param_table);
 
-        for inst in &mut res.data {
-            match &mut inst.op {
-                OperationType::Standard(_) => {
-                    #[cfg(feature = "cache_pygates")]
-                    {
-                        inst.py_op = None;
+        if copy_instructions {
+            for inst in &mut res.data {
+                match &mut inst.op {
+                    OperationType::Standard(_) => {
+                        #[cfg(feature = "cache_pygates")]
+                        {
+                            inst.py_op = None;
+                        }
                     }
-                }
-                OperationType::Gate(ref mut op) => {
-                    op.gate = op.gate.call_method0(py, intern!(py, "copy"))?;
-                    #[cfg(feature = "cache_pygates")]
-                    {
-                        inst.py_op = None;
+                    OperationType::Gate(ref mut op) => {
+                        op.gate = op.gate.call_method0(py, intern!(py, "copy"))?;
+                        #[cfg(feature = "cache_pygates")]
+                        {
+                            inst.py_op = None;
+                        }
                     }
-                }
-                OperationType::Instruction(ref mut op) => {
-                    op.instruction = op.instruction.call_method0(py, intern!(py, "copy"))?;
-                    #[cfg(feature = "cache_pygates")]
-                    {
-                        inst.py_op = None;
+                    OperationType::Instruction(ref mut op) => {
+                        op.instruction = op.instruction.call_method0(py, intern!(py, "copy"))?;
+                        #[cfg(feature = "cache_pygates")]
+                        {
+                            inst.py_op = None;
+                        }
                     }
-                }
-                OperationType::Operation(ref mut op) => {
-                    op.operation = op.operation.call_method0(py, intern!(py, "copy"))?;
-                    #[cfg(feature = "cache_pygates")]
-                    {
-                        inst.py_op = None;
+                    OperationType::Operation(ref mut op) => {
+                        op.operation = op.operation.call_method0(py, intern!(py, "copy"))?;
+                        #[cfg(feature = "cache_pygates")]
+                        {
+                            inst.py_op = None;
+                        }
                     }
-                }
-            };
+                };
+            }
         }
         Ok(res)
     }