@mtreinish 2024-05-25
Oh, those neko failures are interesting and unexpected (I was expecting the 6 unit test failures in terra's unit tests). Looking at it I think this is the same case of something I hit in the nlocal parameter tests. The underlying issue is that for the type of the elements of param vector in rust I defined basically the following type:

```rust
#[derive(FromPyObject)]
enum Param {
    Float(f64),
    ParameterExpression(PyObject)
}
```

and the issue is the derive `FromPyObject` implementation is eagerly trying to cast the input as a float first to determine if it should match that arm. When someone defines a `ParameterExpression` like:

```python
p = Parameter('p')
expression = p * 0
```

and then uses `expression` as a gate parameter, the float check will simplify the expression at circuit insertion time instead of preserving the parameter expression. Prior to this PR that normalization to a float would occur only after binding and not at insertion time. I hit this exact case in the nlocal tests and just worked around it by changing the test to be `1 * param` instead of multiplying by zero because I figured it was a pathological edge case and arguably a test bug, because in practice nobody would actually do that, and if they did they wouldn't care that `parm` disappeared because it was meaningless. But, it appears that the neko tests are doing this too so maybe it's better to fix this for real. The best way would probably to stop deriving `FromPyObject` and manually implement it for `Param` to first check if the input is a `ParameterExpression`, then try a float, lastly fall back to a `PyObject` (this is necessary for custom instruction types that define their own parameter types, `UnitaryGate` being an example). I guess this has the advantage of being able to distinguish between a `ParameterExpression` and a generic object at insertion time which will simplify the parameter table logic (which was doing this differentiation explicitly).