@mtreinish 2024-05-28
The ci failure is the docs build failing because of a warning on import from `qiskit-qasm3-import` because `qiskit-qasm3-import` was using private attributes and classes that are removed in this PR, see: https://github.com/Qiskit/qiskit-qasm3-import/issues/33.