@jlapeyre 2024-05-29
Here is a minimal example that fails stochastically for me. For example on https://github.com/mtreinish/qiskit-core/commit/dfb02de7246c92425d2a6f765e128c09cb9e71dc

```rust
from qiskit import QuantumRegister, QuantumCircuit
from qiskit.qasm3 import Exporter, dumps

def test_no_include():
    """Test explicit gate declaration (no include)"""
    q = QuantumRegister(2, "q")
    circuit = QuantumCircuit(q)
    circuit.cx(0, 1)
    circuit.sx(0)
    Exporter(includes=[]).dumps(circuit).splitlines()
```

You have to compile with `QISKIT_NO_CACHE_GATES=1`