@eliarbel 2024-06-03
I was wandering if there is a cleaner way to handle the mapping from the Python gate objects to their Rust representation, from the Python API perspective. Specifically, I'm addressing the use of the `_standard_gate` class attribute in all the Rust-supported gate classes. 

Have you considered a more centralized solution? For example, something like a translation table that maps a give Python gate instance to its Rust type or some callable? This can promote more decoupling between the Python and Rust parts (which I think is desirable) and keep the Python API cleaner. 

I was playing with a naive code just to gauge performance of one approach. It seems not so bad, but again, this might be too naive: 
```python
for i in range(100):
    exec(f"class GATE{i}: pass") # creating mock gate classes 

translation_table = {}
gates = list()
for i in range(100):
    exec(f"translation_table[type(GATE{i}())] = {i}")  # This is a translation table from gate type to some useful data
    exec(f"gates.append(GATE{i}())") # to be used later
```

and then: 
```python
%%timeit
translation_table[type(gates[random.randint(0,99)])]
```
771 ns ± 54.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

where: 
```python
%%timeit 
random.randint(0,99)
```
572 ns ± 37.8 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)