@mtreinish 2024-06-03
> Have you considered a more centralized solution? For example, something like a translation table that maps a given Python gate instance to its Rust type or some callable? This can promote more decoupling between the Python and Rust parts (which I think is desirable) and keep the Python API cleaner.

This was actually the first thing I tried. There are 2 issues, the first which was my primary motivator, is the import tree precludes doing this because we'll hit import cycles when loading gate classes as the rust code will need to import things more eagerly to build the mapping. There wasn't a simple path to resolve the cycles that I could find. The second issue is the performance, just checking if there is an attribute and it's value is `None` is significantly faster and easier via the C API we call from rust. We're going to be checking every inbound object on the circuit if it's a standard gate, so the efficiency will be quite significant in aggregate.