@eliarbel 2024-06-03
Thanks Matthew. 2 follow up comments: 
> There are 2 issues, the first which was my primary motivator, is the import tree precludes doing this because we'll hit import cycles when loading gate classes as the rust code will need to import things more eagerly to build the mapping. 

The translation can be kept on the Python side, say maybe in `QuantumCircuit._append_standard_gate()`. This way all the gate classes stay clean and there is only one place to maintain the mapping. Would that work conceptually? 

>The second issue is the performance, just checking if there is an attribute and it's value is None is significantly faster and easier via the C API we call from rust. We're going to be checking every inbound object on the circuit if it's a standard gate, so the efficiency will be quite significant in aggregate.

Agree, a super important concern. But if it the mapping is on the Python side and occurs only once for each append call, it seems the overhead might be reasonable.