@mtreinish 2024-06-03
I've been running some performance benchmarks on this, this definitely causes a large (~80% to 100% slowdown in my benchmarking so far) regression in transpilation because of the overhead of the rust<->python conversion we have to do every time we add access a gate in the circuit. This will be fixed as we start to implement: https://github.com/Qiskit/qiskit/issues/12208

But the circuit construction numbers are looking quite good:

```
Benchmarks that have improved:

| Change   | Before [9d03b4bf] <main>   | After [9e211162] <gates-in-rust>   |   Ratio | Benchmark (Parameter)                                                               |
|----------|-----------------------------------------|------------------------------------|---------|-------------------------------------------------------------------------------------|
| -        | 17.5±0.04μs                             | 15.4±0.1μs                         |    0.88 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 8)             |
| -        | 21.3±0.09μs                             | 18.7±0.2μs                         |    0.88 | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 8)             |
| -        | 14.0±0.06μs                             | 11.8±0.09μs                        |    0.84 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 8)               |
| -        | 20.0±0.2μs                              | 16.7±0.1μs                         |    0.84 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 8)             |
| -        | 20.6±0.5μs                              | 17.0±0.6μs                         |    0.83 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 8)              |
| -        | 31.1±0.09μs                             | 25.8±0.3μs                         |    0.83 | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 8)             |
| -        | 50.6±0.8μs                              | 40.7±0.6μs                         |    0.8  | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 8)            |
| -        | 10.3±0.05ms                             | 8.07±0.05ms                        |    0.79 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 8192)          |
| -        | 26.7±0.5μs                              | 20.8±0.8μs                         |    0.78 | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 8)              |
| -        | 16.7±0.1μs                              | 13.1±1μs                           |    0.78 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 8)               |
| -        | 2.56±0.01ms                             | 2.01±0.01ms                        |    0.78 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 2048)          |
| -        | 70.0±0.3μs                              | 54.4±0.9μs                         |    0.78 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 8)            |
| -        | 71.3±0.2μs                              | 54.3±0.2μs                         |    0.76 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 8)       |
| -        | 169±0.3μs                               | 127±2μs                            |    0.75 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 128)           |
| -        | 12.9±0.02μs                             | 9.52±0.04μs                        |    0.74 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 8)               |
| -        | 40.9±0.1ms                              | 30.4±0.4ms                         |    0.74 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 32768)         |
| -        | 2.79±0.02ms                             | 2.05±0.02ms                        |    0.74 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 2048)         |
| -        | 2.80±0.01ms                             | 2.07±0.03ms                        |    0.74 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 2048)         |
| -        | 11.4±0.09ms                             | 8.32±0.1ms                         |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 2048)    |
| -        | 45.5±0.6ms                              | 33.1±0.3ms                         |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 8192)    |
| -        | 98.7±0.2μs                              | 71.7±0.4μs                         |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 8)       |
| -        | 13.4±0.1μs                              | 9.78±0.06μs                        |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 8)               |
| -        | 199±3μs                                 | 146±4μs                            |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 128)          |
| -        | 2.73±0ms                                | 2.01±0.01ms                        |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 2048)          |
| -        | 229±2μs                                 | 168±3μs                            |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 128)          |
| -        | 2.79±0.03ms                             | 2.04±0.03ms                        |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 2048)          |
| -        | 2.83±0.01ms                             | 2.07±0.03ms                        |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 2048)          |
| -        | 760±4μs                                 | 547±6μs                            |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 128)     |
| -        | 245±0.7μs                               | 176±1μs                            |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 8)      |
| -        | 90.8±0.2μs                              | 65.7±0.4μs                         |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 8)       |
| -        | 148±0.3μs                               | 107±0.5μs                          |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 8)       |
| -        | 40.0±0.9μs                              | 28.9±0.7μs                         |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 128)            |
| -        | 165±0.8ms                               | 119±0.7ms                          |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 131072)        |
| -        | 11.1±0.1ms                              | 7.98±0.06ms                        |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 8192)         |
| -        | 181±0.7μs                               | 131±3μs                            |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 128)           |
| -        | 11.1±0.05ms                             | 7.97±0.03ms                        |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 8192)         |
| -        | 196±0.7μs                               | 142±3μs                            |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 128)           |
| -        | 197±0.8μs                               | 142±3μs                            |    0.72 | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 128)           |
| -        | 716±3ms                                 | 509±3ms                            |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 131072)  |
| -        | 182±1ms                                 | 128±1ms                            |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 32768)   |
| -        | 339±0.9μs                               | 242±1μs                            |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 8)      |
| -        | 47.4±0.7μs                              | 33.8±0.8μs                         |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 128)            |
| -        | 36.0±0.1μs                              | 25.7±0.07μs                        |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 128)             |
| -        | 11.2±0.03ms                             | 7.91±0.1ms                         |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 8192)          |
| -        | 11.3±0.03ms                             | 8.04±0.04ms                        |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 8192)          |
| -        | 11.0±0.1ms                              | 7.77±0.05ms                        |    0.7  | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 8192)          |
| -        | 177±3ms                                 | 122±2ms                            |    0.69 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 131072)       |
| -        | 174±0.9ms                               | 120±2ms                            |    0.69 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 131072)        |
| -        | 44.2±0.4ms                              | 30.1±0.6ms                         |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 32768)        |
| -        | 44.4±0.4ms                              | 30.1±0.2ms                         |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 32768)         |
| -        | 44.5±0.1ms                              | 30.2±0.4ms                         |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 32768)        |
| -        | 44.9±0.1ms                              | 30.5±0.1ms                         |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 32768)         |
| -        | 178±1ms                                 | 121±0.5ms                          |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 131072)        |
| -        | 44.6±0.2ms                              | 30.3±0.2ms                         |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 32768)         |
| -        | 970±20μs                                | 651±5μs                            |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 128)    |
| -        | 52.7±0.6ms                              | 35.2±0.2ms                         |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 8192)   |
| -        | 12.9±0.08ms                             | 8.63±0.04ms                        |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 2048)    |
| -        | 51.0±0.9ms                              | 34.3±0.5ms                         |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 8192)    |
| -        | 1.11±0ms                                | 741±4μs                            |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 128)    |
| -        | 178±1ms                                 | 120±0.8ms                          |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 131072)       |
| -        | 178±0.6ms                               | 120±0.9ms                          |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 131072)        |
| -        | 13.3±0.06ms                             | 8.83±0.02ms                        |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 2048)   |
| -        | 210±2ms                                 | 139±2ms                            |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 32768)  |
| -        | 891±9μs                                 | 585±5μs                            |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 128)     |
| -        | 812±7ms                                 | 536±4ms                            |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 131072)  |
| -        | 203±4ms                                 | 134±1ms                            |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 32768)   |
| -        | 13.3±0.05ms                             | 8.74±0.07ms                        |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 2048)   |
| -        | 947±4μs                                 | 623±7μs                            |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 128)     |
| -        | 13.3±0.09ms                             | 8.86±0.06ms                        |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 2048)    |
| -        | 52.5±0.8ms                              | 34.8±0.6ms                         |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 8192)    |
| -        | 833±5ms                                 | 540±2ms                            |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 131072) |
| -        | 209±1ms                                 | 137±1ms                            |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 32768)  |
| -        | 52.5±0.2ms                              | 34.3±0.3ms                         |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 8192)   |
| -        | 838±4ms                                 | 542±0.9ms                          |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 131072)  |
| -        | 208±0.7ms                               | 136±2ms                            |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 32768)   |
| -        | 960±4μs                                 | 619±4μs                            |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 128)     |
| -        | 13.3±0.02ms                             | 8.63±0.2ms                         |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 2048)    |
| -        | 211±1ms                                 | 137±2ms                            |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 32768)   |
| -        | 52.8±0.2ms                              | 34.4±0.6ms                         |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 8192)    |
| -        | 843±7ms                                 | 540±2ms                            |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 131072) |
| -        | 843±7ms                                 | 543±2ms                            |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 131072)  |
| -        | 361±5μs                                 | 231±80μs                           |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 2048)            |
| -        | 1.42±0.02ms                             | 908±300μs                          |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 8192)            |
| -        | 366±4μs                                 | 233±4μs                            |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 2048)            |
| -        | 366±7μs                                 | 232±70μs                           |    0.63 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 2048)           |
| -        | 368±6μs                                 | 228±0.6μs                          |    0.62 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 2048)            |
| -        | 371±6μs                                 | 232±80μs                           |    0.62 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 2048)            |
| -        | 375±6μs                                 | 229±70μs                           |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 2048)           |
| -        | 1.44±0.03ms                             | 858±20μs                           |    0.6  | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 8192)            |
| -        | 1.43±0.03ms                             | 862±20μs                           |    0.6  | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 8192)            |
| -        | 1.45±0.01ms                             | 869±20μs                           |    0.6  | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 8192)            |
| -        | 37.9±0.4μs                              | 16.8±0.8μs                         |    0.44 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 128)             |
| -        | 33.4±0.5μs                              | 13.8±0.05μs                        |    0.41 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 128)             |
| -        | 5.62±0.08ms                             | 2.32±0.03ms                        |    0.41 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 32768)           |
| -        | 33.6±0.5μs                              | 13.4±0.1μs                         |    0.4  | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 128)             |
| -        | 5.75±0.07ms                             | 2.22±0.05ms                        |    0.39 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 32768)           |
| -        | 5.70±0.08ms                             | 2.14±0.5ms                         |    0.38 | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 32768)          |
| -        | 5.68±0.09ms                             | 2.17±0.06ms                        |    0.38 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 32768)           |
| -        | 21.7±0.2ms                              | 7.94±0.08ms                        |    0.37 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 131072)          |
| -        | 5.75±0.1ms                              | 2.14±0.07ms                        |    0.37 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 32768)          |
| -        | 22.1±0.4ms                              | 7.31±0.08ms                        |    0.33 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 131072)         |
…